fix: 사용일수 조회 로직 수정

+++ src/main/java/com/nanal/backend/domain/mypage/dto/resp/RespGetServiceLife.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Data
 public class RespGetServiceLife {
 
-    private Integer serviceLife;
+    private long serviceLife;
 }
